billing: move the monthly run into its own module
